+++ CHANGELOG.md
@@ -6,6 +6,18 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.1.0/),
 
 *Note*: We try to adhere to these practices as of version [v0.2.1].
 
+## Version [0.3.0] - 2024-06-8
+
+### Changed
+
+- adapted the LaplaceClassification and the LaplaceRegression struct to use the new @mlj_model macro from MLJBase.
+- Changed the fit! method arguments. Now it also accept a Flux chain model instead of retrieving it from the structs. this is due to the fact that MLJ wants only hyperparameters in the struct https://juliaai.github.io/MLJModelInterface.jl/dev/quick_start_guide/
+- Changed the predict functions for both LaplaceClassification and  LaplaceRegression.
+
+### Removed
+- Removed the shape , build and clean! functions.
+
+
 ## Version [0.2.3] - 2024-05-31
 
 ### Changed